+++ apps/desktop/src/state/operations/agent-orchestrator/events/session-helpers.ts
@@ -1,7 +1,8 @@
 import type { AgentSessionState } from "@/types/agent-orchestrator";
 import { isTodoToolName, settleDanglingTodoToolMessages } from "../../agent-tool-messages";
+import { finalizeDraftAssistantMessage } from "../support/assistant-meta";
 import { runOrchestratorSideEffect } from "../support/async-side-effects";
-import { finalizeDraftAssistantMessage, sanitizeStreamingText } from "../support/utils";
+import { sanitizeStreamingText } from "../support/core";
 import type {
   DraftChannel,
   DraftChannelValueMap,
